Syntax: Introduce lexical scoping
Define a class for representing lexical scopes, used during compile
time. These differ from the environment objects used at runtime, which
are instantiated every time a scope defining expression (such as letrec
or the body of a lambda procedure) is evaluated.

Every lexical scope has an array of slots, corresponding to variable
bindings that will be in place at runtime. Symbol resolution is
performed at compile time, producing a reference consisting of (a) the
number of times we must move up the chain of environment objects to find
the one containing the target variable value, and (b) the index of the
value within the array of variable values maintained by the environment
object.

This commit also introduces the set! special form, which can be used to
assign new values to variables at runtime.