CPS eval: Basic continuation-passing evaluation
Add an evaluate() method to each AST node class which evaluates the
expression and then calls either a success or failure continuation
function. The latter are used for exception handling; if an error occurs
during evaluation, or a throw expression is evaluated, the failure
continuation is called with a value corresponding to the exception.

This implementation uses JavaScript's built-in stack, and thus is not
properly tail-recursive. Running non-trivial programs will quickly
produce a stack overflow.