Dataflow: Create classes for each type of node
Introduce a new mode evaluation, activated by --eval-reactive, which
constructs a set of dataflow nodes during evaluation. At present, these
simply wrap the values that are created, and are discarded as soon as
the values are extracted from them after creation. Each DataflowNode
subclass corresponds to an ASTNode class, and in its constructor,
carries out the necessary operations to compute the value of that node
in the given environment. Exception handling (try and throw) forms are
not supported.

While the current use of these classes has the same observable behaviour
as that of direct evaluation, the stage is now set for us to maintain a
graph of dataflow nodes in memory and re-evaluate the necessary parts of
that graph when the values of input variables change.