Dataflow: Make environments bind nodes, not values
Change the bindings in an Environment from a mapping between names and
values to a mapping between names and dataflow nodes. Assignment
expressions cause the value to be bound to a different node, not just a
different value, as we want subsequent references to a given name to
refer to the part of the dataflow graph that corresponds to that name at
a given point during evaluation.

In implementation terms, this means replacing the value field of the
Variable class with a node field. For convenience and to avoid having to
alter the direct and CPS evaluation logic in other parts of the code, we
provide a getter and setter for the value property which defers to the
value property of the node. For these modes the dataflow node object
will simply be a placeholder; for dataflow evaluation the value property