Dataflow: Add input special form
Add a mapping of names to values, referred to as "inputs", which allow
the runtime environment to set and change values in the dataflow graph
from which the program can compute further values. An input can be
retrieved by a new form of expression, (input x), which retrieves the
input named "x" from the mapping.

Currently this is just a static mapping, with no mechanism to handle
changes. Before starting evaluation in reactive mode, we initialise a
single input called "test" to the value 0. In the future, we will re-
evaluate the necessary portions of the dataflow graph when an input, and
any values derived from it, change.