Dataflow: Add change propagation logic
Keep track of the dependencies between different nodes in the dataflow
graph. Each node has an array of inputs and outputs; the inputs are
those other nodes on which this one depends, and the outputs are other
nodes that depend on this one. The addOutput() and removeOutput()
methods maintain these two arrays and are the only means by which they
should be modified.

Note that the term "input" here is distinct from its usage as a location
whose value can be changed by the runtime system externally from the
program itself.

Rudimentary tracing facilities are included to report which nodes are
changed, so that change propagation can be observed. Each dataflow node
is given a unique identifier for use in this and future debugging
facilities.

The DataflowNode class now has two abstract methods, reevaluate() and
detach(). The former is called when the value of one or more inputs has
changed, while the latter is intended to be called when it is known that
the node is no longer needed and will be removed from the graph.