Dataflow: Show changes when printing values
Assign each value a "generation", which allows us to determine which
evaluation cycle it was created in. When converting values to a string,
optionally allow those with a generation equal to or later than that
specified as a parameter to be highlighted in blue. This allows us to
see which values have changed after the dataflow graph has been updated.

Note that the re-evaluation logic in the dataflow nodes is not yet
implemented; this commit just adds the necessary debugging facilities.

Additionally, we sightly simplify error reporting in main.ts.