Dataflow: Repeatedly change inputs & re-evaluate
When running a program with --eval-reactive, increase the counter value
stored in the "test" input once per second, and trigger re-evaluation of
the regular parts of the dataflow graph. At each step, we report which
nodes change their values, and dump the structure of the dataflow graph
to standard output. The latest result value is printed with changes
highlighted.