Dataflow: Re-evaluate apply expressions
Implement re-evaluation logic for ApplyNode.  We cater for both types of
procedures (built-in functions and lambdas), as well as the case where
the a different procedure must be applied due to a change in the value
that appears as the first element of the apply form.

For built-in procedures, we create a BuiltinCallDataflowNode object.
This lists all of the arguments as dependencies, and is re-evaluated
when any of them changes.  This registers itself as an output of all its
arguments, causing it to be re-evaluated if any of them changes.

When re-evaluating, we call the built-in function again, passing a
reference to the node along with the call in case the function mutates a
value. In most cases, such as for arithmetic operatirs, this will not be
the case, and we simply store the new value. For cons specifically, we
keep the same cons cell but update its car and cdr fields. This causes
list structures to be maintained if some of their elements change,
rather than having to re-create the whole list.

For lambda procedures, we instantiate the body in a similar manner to
what would be done in graph reduction (although we do not subsequently
reduce the graph; the body is kept in memory to facilitate dataflow
evaluation). Any data dependencies that expressions in the body have on
the arguments are taken care of as part of this instantiation process,
by looking up the relevant variables in the environment. For this
reason, we do not need a LambdaCallDatafowNode class that would play a
role similar to BuiltinCallDataflowNode. Instead, the body of the
instantiated lambda is returned.

The ApplyDataflowNode contains a reference to the node containing the
procedure value, in case it changes to something else as a result of
dataflow evaulation. When this happens, we repeat the steps above
depending on the type of the new procedure, which alters the structure
of the graph. As long as the procedure does not change (something we may
be able to detect statically in the future), the graph structure remains
fixed, with the exception of any descendant nodes (e.g. if) which may
change.