Dataflow: Re-evaluate if expressions
As with apply expressions, it is necessary to modify the structure of
the dataflow graph for if expressions if the condition changes.

Initially, we create a node for the condition, and use its initial value
to create a "branch" node from either the consequent or alternative,
depending on the initial value of the condition.

When re-evaluating an if expression, we have to check if the condition
has changed; if so, we replace the branch with a newly created subgraph
from either the consequent or alternative, depending on the new value.
If the condition has not changed, then we simply propagate the value
from the branch node to the if node itself.