Dataflow: Make assignment rebind node, not value
When an assignment expression is initially evaluated (that is, when the
corresponding dataflow node is created), modify the environment such
that the specified variable is changed to point to the dataflow node of
the expression being assigned from, rather than its value. Unlike most
other expressions, assignment alters the binding of a name in the
environment, rather than computing a value. We need to make sure that
variable references always point to the correct node.

The current solution does not address the case where the structure of
the graph is changed - for example when an if expression's condition
changes, which causes an alternate branch to be evaluated in which a
different set of assignments is carried out. At present we only properly
handle assignments that are made in a manner that can be determined
statically. In the even that there are any conditionally-executed
assignmets, the bindings will be those that were in place when the first
evaluation occurred.