Syntax: Make ASTNode a discriminated union type
Rename ASTNode to ASTBaseNode, and introduce a new type called ASTNode
which is a union of all the ASTBaseNode subclasses. Add a 'kind' field
to each which can be used to distinguish between different types of AST
nodes.

This allows us to write code in a procedural style where we use a switch
statement to handle each node type appropriately, instead of relying
solely on inherited methods defined on the AST node classes themselves.