Eval: Use procedural instead of OO style
Rather than using overridden methods for evaluating AST nodes (in both
direct and CPS mode), move all of the evaluation logic for each mode
into a single function that deals with all node types. We now have
functions called evalDirect() and evalCps(), defined in their own files,
which take a node object as a parameter and return the result or call
the appropriate continuation, respectively.

I consider this approach more appropriate for the codebase in its
current stage, where we are still experimenting evaluation modes, since
doing so is simpler if we can keep each in a separate file. This is
better than adding an extra method to each AST node class, even if all
that method does is to create another object to perform the evaluation
like we still do for the reactive (dataflow) evaluation mode.

TypeScript's handling of union types in switch statements ensures that
for functions that must return a value, we will get a warning if we add
a new node type that is not handled. This is presently the case for
direct evaluation only; CPS evaluation returns void, and thus throws an
exception if it encounters a node type it doesn't know how to deal with.