Syntax: Hide input data structure from parser
Add two new classes, TextBuffer and Cursor, which hide from the parser
the underlying representation used to store the source code in memory.
The parser can use the cursor to save the current position, advance to
the next character, and can use the text buffer to extract a string
containing the characters from the buffer that lie between two different
positions.

This is a step towards achieving IDE integration where we are notified
of ranges in the text buffer that have changed, and re-parse only the
modified parts of the source rather than the whole file.