Tracing: Initial version of tracing evaluation
Add a new evaluation mode that operates like direct evaluation, except
that it records the tree of function calls that were made. Once the
final result is obtained, this tree is printed to standard output.

For example, running the following code with the --eval-tracing option:

    ((lambda (x y)
       (sqrt (+ (* x x) (* y y))))
     3 4)

Will produce the following output:

    #0 apply
    ├── #1 lambda
    ├── #2 constant 3
    ├── #3 constant 4
    └── #4 apply
        ├── #5 variable sqrt
        └── #6 apply
            ├── #7 variable +
            ├── #8 apply
            │   ├── #9 variable *
            │   ├── #10 variable x
            │   └── #11 variable x
            └── #12 apply
                ├── #13 variable *
                ├── #14 variable y
                └── #15 variable y
    TRACING Success: 5