Tracing: Perform evaluation in sublcasses of Trace
Create subclasses of Trace for each type of syntax construct, and move
the logic for evauating expressions and creating the appropriate child
trace objects into these classes.

Each trace object has a reference to its AST node and the environment in
which it must be evaluated. For all cases within evalTracing(), we first
construct the object and then call evaluate() on it. This results in the
same observable behaviour; the actions are just performed in a different
part of the code.