Tracing: Separate creation & evaluation of Traces
Instead of creating and evaluating the tree of trace objects as part of
the same process, separate the two activities. This is specifically to
allow the evaluate() method to be called multiple times on an already-
existing tree, which will go through the full evaluation process. After
the first iteration, subsequent evaluations will produce exactly the
same result value as the first, *unless* one or more input values used
by the program are changed, e.g. by including an expression of the form
"(input test)" in the code.

Some parts of the tree are not actually constructed initially, and the
structure of the tree may change in cases where the parts cannot be
determined until evaluation. For example, an assignment or sequence
expression always has the same structure, but an if statement does not,
as the branch chosen depends on the conditional value which cannot be
determined statically. Modifying an input to the program and calling
evaluate() on the tree may cause the structure to change, in addition to
returning a different value.